projects/App-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
algosdk==2.10.0
algokit-utils==4.2.0
google-genai==0.3.0
google-generativeai==0.3.2
